+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [Unreleased]
+### Breaking changes
+- ComponentVerticle.registerSelfConfChangeListener accepts Handler<ConfigChanged> instead of Handler<JsonObject>
+
 ## [1.4.0] - 2020-12-30
 ### Added
 - Support for escape character \\ for dots in path in json extractor